ARTA - Adaptive Rhythmic Training Application
Application used to train with and evaluate rhythmic patterns used in musical training. It also uses a Bilateral LTSM neural network trained to predict the average performance of the next 12 notes typed on the keyboard by the participant. There is no music in the program, just a metronome click that changes tempo according to specific parameters tuned adaptively to enhance the leardning of the letters sequences organized in rhythmic patterns. The whole methodology of this work, the conception an creation of the application and its use to evaluate through 3 experiments are described in my bachelor project (B.Sc. Games Programming), which can be found at the following link: 

https://www.academia.edu/79300882/Conception_of_New_Rhythm_Based_Dexterity_Training_Methods_with_Adaptive_Mechanics_and_AI