Unity Environment Example
The executable is at this path:  /Simulation/Simulation.exe

This is a procedurally generated landscape where the player runs away from NPC characters and try to eat all the chocolate bars in the labyrinth, which is also produced differently every time with a recursive algorithm that fills the whole bunker space with ciment blocks. Have fun! In this projects, you can see the basic folder structure and all the scripts used for this simulation.